+++ ProtoDock.QuickLaunch/QuickLaunchIcon.cs
@@ -11,7 +11,7 @@ namespace ProtoDock.QuickLaunch
         public IDockPanelMediator Mediator { get; }
         
         public string Title => _path;
-        public int Width => 1;
+        public float Width => 1;
         public bool Hovered => true;
         
         private string _path;
